requirements.txt
flask
bs4
requests
pymongo
dnspython
python-dotenv
urllib3==1.26.6